stm32l4-multi: extract common RCC code
Extract generic reset and clock controller code for STM32L4 and N6.
